+++ src/main/java/com/example/demo/dto/stock/response/CreateStockResponse.java
@@ -11,7 +11,7 @@ public class CreateStockResponse {
 
     private final Long stockId;
     private final String stockName;
-    private final Double price;
+    private final Long price;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private final LocalDateTime createdAt;
